pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rjson"
version = "0.1.0"
description = "Small YAML-to-JSON template runtime"
authors = [ {name = "dev", email = "dev@example.com"} ]
license = "MIT"
dependencies = ["PyYAML>=6.0", "Faker>=18.0.0"]

[project.scripts]
rjson = "rjson.cli:main"
